Move graph to own folder
