+++ archlens-in-vscode/src/extension.ts
@@ -2,7 +2,7 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 import { getWebviewContent } from './utilities/getWebviewContent';
-import * as graph_util from "./utilities/graph";
+import * as graph_util from "./graph/graph";
 import { showTreeView } from './utilities/treeView';
 import * as archlens from './archlens/archLens';
