+++ archlens-in-vscode/src/utilities/getWebviewContent.ts
@@ -5,8 +5,8 @@ export function getWebviewContent(webview: vscode.Webview, extensionUri: vscode.
     const indexPath = vscode.Uri.joinPath(extensionUri, '..' ,'webview','index.html');
 
     let html = fs.readFileSync(indexPath.fsPath, 'utf8');
-    // Replace placeholders with actual URIs
 
+    // Replace placeholders with actual URIs
     html = html.replace("{baseUri}", webview.asWebviewUri(vscode.Uri.joinPath(extensionUri, '..' ,'webview')).toString());
 
     return html;
